requirements.txt
matplotlib
numpy
fuzzywuzzy
python-Levenshtein
gensim
seaborn
graphviz
pandas
pattern
pytorch-pretrained-bert
Unidecode
sentence-transformers
tensorflow
sklearn
spacy